format: improve CSV output
Properties are now divided into numeric and text properties; the
former are always in decimal and never quoted, the latter are always
quoted.  In addition memory/io properties are always given in byte and
time intervals in seconds.  This should make it easier to get some use
out of CSV output when imported into a spreadsheet.